Remove LessonVideoUpload component from admin panel and move it to admin/pages/Lesson
Add new route for admin dashboard with AdminMain component as default route and LessonVideoUpload as nested route

Remove direct route for LessonVideoUpload and add it as nested route under admin dashboard
